requirements.txt
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.26.2
pyttsx3==2.90
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.3
customtkinter==5.2.1
pillow==10.1.0
python-vlc==3.0.20123